+++ test/tickets/T0183.hs
@@ -1,0 +1,6 @@
+-- SYNTAX TEST "source.haskell" "Character literals above U+007E"
+
+
+let cs =
+    [ '¹', 'á', '¶', '«', '»', '¿', '§', '£' ]
+--     ^    ^    ^    ^    ^    ^    ^    ^ string.quoted.single.haskell