+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - Fix highlighting of data types with fields that contain constraints
   ([#179](https://github.com/JustusAdam/language-haskell/issues/179)).
 - Updated the list of known LANGUAGE extensions to be current as of GHC 10.0.
+- Fix parsing of character literals with Unicode code point above U+007E,
+  such as `'à'`, `'£'`... ([#183](https://github.com/JustusAdam/language-haskell/issues/183)).
 
 ## 3.6.0 - 15.03.2022